+++ packages/integration-test/importDefault.test.mjs
@@ -1,3 +1,5 @@
+/** @jest-environment @happy-dom/jest-environment */
+
 import { act, fireEvent, render, screen } from '@testing-library/react';
 import React from 'react';
 import { DictateButton } from 'react-dictate-button';