+++ packages/react-dictate-button/__tests__/simple.checkbox.spec.tsx
@@ -2,7 +2,7 @@
 
 import { act, fireEvent, render, screen } from '@testing-library/react';
 import React from 'react';
-import { DictateButton, type DictateEventHandler, type EndEventHandler, type StartEventHandler } from '../src/index';
+import { DictateCheckbox, type DictateEventHandler, type EndEventHandler, type StartEventHandler } from '../src/index';
 import {
   SpeechRecognition,
   SpeechRecognitionAlternative,
@@ -11,7 +11,7 @@ import {
   SpeechRecognitionResultList
 } from '../src/internal';
 
-describe('simple scenario', () => {
+describe('simple scenario for <DictateCheckbox>', () => {
   let constructSpeechRecognition: jest.Mock<SpeechRecognition, []>;
   let onDictate: jest.Mock<ReturnType<DictateEventHandler>, Parameters<DictateEventHandler>, undefined>;
   let onEnd: jest.Mock<ReturnType<EndEventHandler>, Parameters<EndEventHandler>, undefined>;
@@ -32,7 +32,7 @@ describe('simple scenario', () => {
     onStart = jest.fn();
 
     render(
-      <DictateButton
+      <DictateCheckbox
         onDictate={onDictate}
         onEnd={onEnd}
         onStart={onStart}
@@ -40,16 +40,19 @@ describe('simple scenario', () => {
         speechRecognition={constructSpeechRecognition}
       >
         Click me
-      </DictateButton>
+      </DictateCheckbox>
     );
   });
 
-  describe('when the dictate button is clicked', () => {
+  describe('when the dictate checkbox is checked', () => {
     beforeEach(() => {
       expect(constructSpeechRecognition).toHaveBeenCalledTimes(0);
 
       act(() => fireEvent.click(screen.getByText('Click me')));
     });
+
+    test('should be checked', () =>
+      expect(screen.getByText('Click me').querySelector('input')).toHaveProperty('checked', true));
 
     test('SpeechRecognition object should be constructed', () =>
       expect(constructSpeechRecognition).toHaveBeenCalledTimes(1));
@@ -71,10 +74,10 @@ describe('simple scenario', () => {
         })
       );
 
-      test('onStart() should have been called once', () => expect(onStart).toHaveBeenCalledTimes(1));
+      test('onStart() should be called', () => expect(onStart).toHaveBeenCalledTimes(1));
       test('onEnd() should not be called', () => expect(onEnd).toHaveBeenCalledTimes(0));
 
-      describe('when result events are dispatched', () => {
+      describe('when result event is dispatched', () => {
         beforeEach(() =>
           act(() => {
             speechRecognition.dispatchEvent(
@@ -110,7 +113,10 @@ describe('simple scenario', () => {
           );
 
           test('onStart() should not be called again', () => expect(onStart).toHaveBeenCalledTimes(1));
-          test('onEnd() should have been called once', () => expect(onEnd).toHaveBeenCalledTimes(1));
+          test('onEnd() should be called', () => expect(onEnd).toHaveBeenCalledTimes(1));
+
+          test('should be unchecked', () =>
+            expect(screen.getByText('Click me').querySelector('input')).toHaveProperty('checked', false));
         });
       });
     });