+++ packages/react-dictate-button/src/Composer.tsx
@@ -5,11 +5,13 @@ import { useRefFrom } from 'use-ref-from';
 
 import Context from './Context.ts';
 import { type DictateEventHandler } from './DictateEventHandler.ts';
+import { type EndEventHandler } from './EndEventHandler.ts';
 import { type ErrorEventHandler } from './ErrorEventHandler.ts';
 import { type ProgressEventHandler } from './ProgressEventHandler.ts';
 import { type RawEventHandler } from './RawEventHandler.ts';
 import { type SpeechGrammarListPolyfill } from './SpeechGrammarListPolyfill.ts';
 import { type SpeechRecognitionPolyfill } from './SpeechRecognitionPolyfill.ts';
+import { type StartEventHandler } from './StartEventHandler.ts';
 import { type TypedEventHandler } from './TypedEventHandler.ts';
 import usePrevious from './usePrevious.ts';
 import vendorPrefix from './vendorPrefix.ts';
@@ -25,13 +27,16 @@ type ComposerProps = {
       ) => ReactNode)
     | ReactNode
     | undefined;
+  continuous?: boolean | undefined;
   extra?: Record<string, unknown> | undefined;
   grammar?: string | undefined;
   lang?: string | undefined;
   onDictate?: DictateEventHandler | undefined;
+  onEnd?: EndEventHandler | undefined;
   onError?: ErrorEventHandler | undefined;
   onProgress?: ProgressEventHandler | undefined;
   onRawEvent?: RawEventHandler | undefined;
+  onStart?: StartEventHandler | undefined;
   speechGrammarList?: SpeechGrammarListPolyfill | undefined;
   speechRecognition?: SpeechRecognitionPolyfill | undefined;
   started?: boolean | undefined;
@@ -59,13 +64,16 @@ function recognitionAbortable(recognition: unknown): recognition is {
 
 const Composer = ({
   children,
+  continuous,
   extra,
   grammar,
   lang,
   onDictate,
+  onEnd,
   onError,
   onProgress,
   onRawEvent,
+  onStart,
   speechGrammarList = navigator.mediaDevices &&
     // @ts-expect-error navigator.mediaDevices.getUserMedia may not be defined in older browsers.
     navigator.mediaDevices.getUserMedia &&
@@ -77,17 +85,22 @@ const Composer = ({
   started
 }: ComposerProps) => {
   const [readyState, setReadyState] = useState(0);
-  const emitDictateOnEndRef = useRef(false);
+  const continuousRef = useRefFrom(continuous);
   const extraRef = useRefFrom(extra);
   const grammarRef = useRefFrom(grammar);
   const langRef = useRefFrom(lang);
   const notAllowedRef = useRef(false);
   const onDictateRef = useRefFrom(onDictate);
+  const onEndRef = useRefFrom(onEnd);
   const onErrorRef = useRefFrom(onError);
   const onProgressRef = useRefFrom(onProgress);
   const onRawEventRef = useRefFrom(onRawEvent);
+  const onStartRef = useRefFrom(onStart);
   const prevSpeechRecognition = usePrevious(speechRecognition);
   const recognitionRef = useRef<SpeechRecognition>();
+  const shouldEmitDictateOnEndRef = useRef(false);
+  const shouldEmitEndRef = useRef(false);
+  const shouldEmitStartRef = useRef(true);
   const speechGrammarListRef = useRefFrom(speechGrammarList);
   const speechRecognitionRef = useRefFrom(speechRecognition);
 
@@ -95,6 +108,24 @@ const Composer = ({
   if (prevSpeechRecognition !== speechRecognition) {
     notAllowedRef.current = false;
   }
+
+  const emitEnd = useCallback(() => {
+    if (shouldEmitEndRef.current) {
+      onEndRef.current?.(new Event('end') as Event & { type: 'end' });
+
+      shouldEmitEndRef.current = false;
+      shouldEmitStartRef.current = true;
+    }
+  }, [onEndRef, shouldEmitEndRef, shouldEmitStartRef]);
+
+  const emitStart = useCallback(() => {
+    if (shouldEmitStartRef.current) {
+      onStartRef.current?.(new Event('start') as Event & { type: 'start' });
+
+      shouldEmitEndRef.current = true;
+      shouldEmitStartRef.current = false;
+    }
+  }, [onStartRef, shouldEmitEndRef, shouldEmitStartRef]);
 
   const handleAudioEnd = useCallback<TypedEventHandler<Event>>(
     ({ target }) => target === recognitionRef.current && setReadyState(3),
@@ -111,10 +142,10 @@ const Composer = ({
 
       // Web Speech API does not emit "result" when nothing is heard, and Chrome does not emit "nomatch" event.
       // Because we emitted onProgress, we should emit "dictate" if not error, so they works in pair.
-      emitDictateOnEndRef.current = true;
+      shouldEmitDictateOnEndRef.current = true;
       onProgressRef.current && onProgressRef.current({ abortable: recognitionAbortable(target), type: 'progress' });
     },
-    [emitDictateOnEndRef, onProgressRef, recognitionRef, setReadyState]
+    [shouldEmitDictateOnEndRef, onProgressRef, recognitionRef, setReadyState]
   );
 
   const handleEnd = useCallback<TypedEventHandler<Event>>(
@@ -123,15 +154,16 @@ const Composer = ({
         return;
       }
 
+      emitEnd();
       recognitionRef.current = undefined;
       setReadyState(0);
 
-      if (emitDictateOnEndRef.current) {
+      if (shouldEmitDictateOnEndRef.current) {
         onDictateRef.current && onDictateRef.current({ type: 'dictate' });
-        emitDictateOnEndRef.current = false;
+        shouldEmitDictateOnEndRef.current = false;
       }
     },
-    [emitDictateOnEndRef, onDictateRef, recognitionRef, setReadyState]
+    [shouldEmitDictateOnEndRef, onDictateRef, recognitionRef, setReadyState]
   );
 
   const handleError = useCallback<TypedEventHandler<SpeechRecognitionErrorEvent>>(
@@ -141,7 +173,7 @@ const Composer = ({
       }
 
       // Error out, no need to emit "dictate"
-      emitDictateOnEndRef.current = false;
+      shouldEmitDictateOnEndRef.current = false;
       recognitionRef.current = undefined;
 
       if (event.error === 'not-allowed') {
@@ -151,8 +183,9 @@ const Composer = ({
       setReadyState(0);
 
       onErrorRef.current && onErrorRef.current(event);
+      emitEnd();
     },
-    [emitDictateOnEndRef, onErrorRef, notAllowedRef, recognitionRef, setReadyState]
+    [emitEnd, onErrorRef, notAllowedRef, recognitionRef, setReadyState, shouldEmitDictateOnEndRef]
   );
 
   const handleRawEvent = useCallback<TypedEventHandler<Event>>(
@@ -167,47 +200,72 @@ const Composer = ({
   );
 
   const handleResult = useCallback<TypedEventHandler<SpeechRecognitionEvent>>(
-    ({ results: rawResults, target }) => {
+    ({ resultIndex, results: rawResults, target }) => {
       if (target !== recognitionRef.current) {
         return;
       }
 
       if (rawResults.length) {
-        const results = Object.freeze(
-          Array.from(rawResults).map(alts => {
-            // Destructuring breaks Angular due to a bug in Zone.js.
-            // eslint-disable-next-line prefer-destructuring
-            const firstAlt = alts[0];
-
-            return {
-              confidence: firstAlt?.confidence || 0,
-              transcript: firstAlt?.transcript || ''
-            };
-          })
-        );
-
         // Destructuring breaks Angular due to a bug in Zone.js.
         // eslint-disable-next-line prefer-destructuring
-        const first = rawResults[0];
+        const rawResult = rawResults[resultIndex];
 
-        if (first?.isFinal) {
-          // After "onDictate" callback, the caller should be able to set "started" to false on an unabortable recognition.
-          recognitionRef.current = undefined;
-          setReadyState(0);
+        if (rawResult?.isFinal) {
+          if (!continuousRef.current) {
+            // After "onDictate" callback, the caller should be able to set "started" to false on an unabortable recognition.
+            emitEnd();
+            recognitionRef.current = undefined;
+            setReadyState(0);
+          }
 
-          onDictateRef.current && onDictateRef.current({ result: results[0], type: 'dictate' });
+          const alt = rawResult[0];
+
+          alt &&
+            onDictateRef.current &&
+            onDictateRef.current({
+              result: {
+                confidence: alt.confidence,
+                transcript: alt.transcript
+              },
+              type: 'dictate'
+            });
+
+          shouldEmitDictateOnEndRef.current = false;
         } else {
-          onProgressRef.current &&
-            onProgressRef.current({ abortable: recognitionAbortable(target), results, type: 'progress' });
+          onProgressRef.current?.({
+            abortable: recognitionAbortable(target),
+            results: Object.freeze(
+              Array.from(rawResults)
+                .filter(result => !result.isFinal)
+                .map(alts => {
+                  // Destructuring breaks Angular due to a bug in Zone.js.
+                  // eslint-disable-next-line prefer-destructuring
+                  const firstAlt = alts[0];
+
+                  return {
+                    confidence: firstAlt?.confidence || 0,
+                    transcript: firstAlt?.transcript || ''
+                  };
+                })
+            ),
+            type: 'progress'
+          });
+
+          shouldEmitDictateOnEndRef.current = true;
         }
       }
     },
-    [onDictateRef, onProgressRef, recognitionRef, setReadyState]
+    [continuousRef, emitEnd, onDictateRef, onProgressRef, recognitionRef, setReadyState, shouldEmitDictateOnEndRef]
   );
 
   const handleStart = useCallback<TypedEventHandler<Event>>(
-    ({ target }) => target === recognitionRef.current && setReadyState(1),
-    [recognitionRef, setReadyState]
+    ({ target }) => {
+      if (target === recognitionRef.current) {
+        emitStart();
+        setReadyState(1);
+      }
+    },
+    [emitStart, recognitionRef, setReadyState]
   );
 
   useEffect(() => {
@@ -229,18 +287,19 @@ const Composer = ({
         recognition.lang = langRef.current;
       }
 
+      recognition.continuous = !!continuous;
       recognition.interimResults = true;
-      recognition.onaudioend = applyAll(handleAudioEnd, handleRawEvent);
-      recognition.onaudiostart = applyAll(handleAudioStart, handleRawEvent);
-      recognition.onend = applyAll(handleEnd, handleRawEvent);
-      recognition.onerror = applyAll(handleError, handleRawEvent);
-      recognition.onnomatch = handleRawEvent;
-      recognition.onresult = applyAll(handleResult, handleRawEvent);
-      recognition.onsoundend = handleRawEvent;
-      recognition.onsoundstart = handleRawEvent;
-      recognition.onspeechend = handleRawEvent;
-      recognition.onspeechstart = handleRawEvent;
-      recognition.onstart = applyAll(handleStart, handleRawEvent);
+      recognition.addEventListener('audiostart', applyAll(handleAudioStart, handleRawEvent));
+      recognition.addEventListener('audioend', applyAll(handleAudioEnd, handleRawEvent));
+      recognition.addEventListener('end', applyAll(handleEnd, handleRawEvent));
+      recognition.addEventListener('error', applyAll(handleError, handleRawEvent));
+      recognition.addEventListener('nomatch', handleRawEvent);
+      recognition.addEventListener('result', applyAll(handleResult, handleRawEvent));
+      recognition.addEventListener('soundend', handleRawEvent);
+      recognition.addEventListener('soundstart', handleRawEvent);
+      recognition.addEventListener('speechend', handleRawEvent);
+      recognition.addEventListener('speechstart', handleRawEvent);
+      recognition.addEventListener('start', applyAll(handleStart, handleRawEvent));
 
       const { current: extra } = extraRef;
 
@@ -259,6 +318,8 @@ const Composer = ({
       const { current: recognition } = recognitionRef;
 
       if (recognition) {
+        emitEnd();
+
         if (recognitionAbortable(recognition)) {
           recognition.abort();
         } else {
@@ -267,6 +328,7 @@ const Composer = ({
       }
     };
   }, [
+    emitEnd,
     extraRef,
     grammarRef,
     handleAudioEnd,