+++ packages/react-dictate-button/src/EndEventHandler.ts
@@ -1,0 +1,5 @@
+import { type TypedEventHandler } from './TypedEventHandler.ts';
+
+export type EndEventHandler = TypedEventHandler<{
+  type: 'end';
+}>;